Fixed bug with add progress (closes #70) and a bit of clean code - fixed code smells found by sonarcloud:   - naming scheme: controller, model etc. lowercase   - removed unnecessary imports   - fixed .* imports   - added javadoc
