+++ src/main/java/tapp/org/tapp/controllers/AthleteController.java
@@ -1,7 +1,5 @@
-package tapp.org.tapp.Controllers;
+package tapp.org.tapp.controllers;
 
-
-import java.util.List;
 
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.criteria.CriteriaBuilder;
@@ -9,14 +7,20 @@ import jakarta.persistence.criteria.CriteriaQuery;
 import jakarta.persistence.criteria.Predicate;
 import jakarta.persistence.criteria.Root;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import tapp.org.tapp.models.Athlete;
+import tapp.org.tapp.repository.AthleteRepository;
 
-
-import tapp.org.tapp.Repository.AthleteRepository;
-import tapp.org.tapp.Models.Athlete;
+import java.util.List;
 
 import static org.springframework.data.jpa.domain.Specification.where;
-import static tapp.org.tapp.Repository.AthleteRepository.*;
+import static tapp.org.tapp.repository.AthleteRepository.isTenantID;
 
 @RestController
 @RequestMapping("/api/v1/")
@@ -36,16 +40,36 @@ public class AthleteController {
 		return athleteRepository.findAll(where(isTenantID(tenantID)));
 	}
 
+	/**
+	 * Add a new athlete to a given tenant
+	 * @param tenantID
+	 * @param athlete JSON Body of the new athlete
+	 * @return JSON of newly added athlete
+	 */
+
 	@PostMapping("/{tenantID}/add_athlete")
 	public Athlete addAthlete(@PathVariable Long tenantID, @RequestBody Athlete athlete) {
 		athlete.setTenantID(tenantID);
 		return athleteRepository.save(athlete);
 	}
 
+	/**
+	 * Update an athlete with the given athleteID and tenantID
+	 * @param tenantID
+	 * @param athleteID
+	 * @param athlete JSON of the updatet athlete data
+	 */
+
 	@PostMapping("/{tenantID}/athletes/{athleteID}")
 	public void updateAthlete(@PathVariable Long tenantID,@PathVariable Long athleteID, @RequestBody Athlete athlete) {
 		athleteRepository.updateAthlete(tenantID, athleteID, athlete.getFirstName(), athlete.getLastName(), athlete.getImageType(),athlete.getPicture());
 	}
+
+	/**
+	 * Delete an athlete with the given athleteID and tenantID
+	 * @param tenantID has to be corrsponding to athlete, otherwise nothing happens
+	 * @param athleteID
+	 */
 
 	@DeleteMapping("/{tenantID}/athletes/{athleteID}")
 	public void deleteAthlete(@PathVariable Long tenantID, @PathVariable Long athleteID){
@@ -56,7 +80,7 @@ public class AthleteController {
 		athleteRepository.deleteById(athleteID);
 	}
 	/**
-	 * Find athlete by athlete ID and tenant ID
+	 * Find athlete by athlete ID and corresponding tenant ID
 	 * @param tenantId
 	 * @param athleteId
 	 * @return Athlete with the given athlete ID
@@ -70,7 +94,7 @@ public class AthleteController {
 	 * Searches athletes with the first or last name containing the search term and the correct tenant ID
 	 * @param tenantID
 	 * @param searchterm
-	 * @return List of athletes with tenantId and searchterm
+	 * @return List of athletes with given tenantId and searchterm
 	 */
 	@GetMapping("/{tenantID}/athletes/{searchterm}")
 	public List<Athlete> findAthletesByName(@PathVariable long tenantID, @PathVariable String searchterm) {
@@ -86,8 +110,7 @@ public class AthleteController {
 		Predicate predicateName = cb.or(predicateLastName, predicateFirstName);
 		Predicate predicateTenantAndName = cb.and(predicateName, predicateTenant);
 		criteriaQuery.where(predicateTenantAndName);
-		List<Athlete> athletes = entityManager.createQuery(criteriaQuery).getResultList();
-		return athletes;
+		return entityManager.createQuery(criteriaQuery).getResultList();
 	}
 }
