+++ src/main/java/tapp/org/tapp/controllers/ProgressController.java
@@ -1,13 +1,18 @@
-package tapp.org.tapp.Controllers;
+package tapp.org.tapp.controllers;
 
 
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.*;
-import tapp.org.tapp.Models.Athlete;
-import tapp.org.tapp.Models.Progress;
-import tapp.org.tapp.Repository.AthleteRepository;
-import tapp.org.tapp.Repository.ProgressRepository;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import tapp.org.tapp.models.Athlete;
+import tapp.org.tapp.models.Progress;
+import tapp.org.tapp.repository.AthleteRepository;
+import tapp.org.tapp.repository.ProgressRepository;
 
 import java.util.List;
 
@@ -24,19 +29,18 @@ public class ProgressController {
 	 * Get all progress entries for all tenants
 	 * @return List of progresses
 	 */
-	// get every progress
 	@GetMapping("/progress")
 	public List<Progress> getEveryProgress(){
 		return progressRepository.findAll();
 	}
 
 	/**
-	 * Get progress for a given athlete
-	 * @param tenantId
+	 * Gets 5 newest progress entries for a given athlete
+	 * @param tenantId of the athlete
 	 * @param athleteId
 	 * @return List of progresses
 	 */
-	@Transactional  // is needed because of the Lob field (comment)
+	@Transactional
 	@GetMapping("/{tenantId}/progress/{athleteId}")
 	public List<Progress> findProgressByAthleteId(@PathVariable long tenantId, @PathVariable long athleteId){
 		if (athleteRepository.getReferenceById(athleteId).getTenant().equals(tenantId)){
@@ -47,11 +51,12 @@ public class ProgressController {
 
 	/**
 	 * Adds progress for a given athlete
-	 * @param tenantId
+	 * @param tenantId of the athlete
 	 * @param athleteId
-	 * @param progress
-	 * @return progress
+	 * @param progress that should be added to the athlete
+	 * @return the newly added progress
 	 */
+	@Transactional
 	@PostMapping("/{tenantId}/progress/{athleteId}")
 	public Progress addProgress(@PathVariable long tenantId, @PathVariable long athleteId, @RequestBody Progress progress) {
 		progress.setAthleteId(athleteId);
@@ -64,15 +69,17 @@ public class ProgressController {
 	/**
 	 * Returns the newest progress entry for each athlete of given tenant
 	 * @param tenantId
-	 * @return List of progresses
+	 * @return List of progresses (One per athelete)
 	 */
-	@Transactional  // is needed because of the Lob field (comment)
+	@Transactional
 	@GetMapping("/{tenantId}/progress/")
 	public List<Progress> findProgressByTenant(@PathVariable Long tenantId){
 		List<Athlete> athletes = athleteRepository.findAthletesByTenantID(tenantId);
 		List<Progress> progressOfTenant = null;
-		for (Athlete athlete: athletes){
-			progressOfTenant.add(progressRepository.findFirstByAthleteIdOrderByCreatedAtDesc(athlete.getId()));
+		if (!athletes.isEmpty()) {
+			for (Athlete athlete : athletes) {
+				progressOfTenant.add(progressRepository.findFirstByAthleteIdOrderByCreatedAtDesc(athlete.getId()));
+			}
 		}
 		return progressOfTenant;
 	}
