+++ src/main/java/tapp/org/tapp/controllers/TenantSkillController.java
@@ -1,10 +1,15 @@
-package tapp.org.tapp.Controllers;
+package tapp.org.tapp.controllers;
 
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.*;
-import tapp.org.tapp.Models.TenantSkill;
-import tapp.org.tapp.Repository.TenantSkillRepository;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import tapp.org.tapp.models.TenantSkill;
+import tapp.org.tapp.repository.TenantSkillRepository;
 
 import java.util.List;
