+++ src/main/java/tapp/org/tapp/models/Athlete.java
@@ -1,9 +1,18 @@
-package tapp.org.tapp.Models;
+package tapp.org.tapp.models;
 
-import java.util.Base64;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Lob;
+import jakarta.persistence.Table;
+
 import java.util.Objects;
 
-import jakarta.persistence.*;
+/**
+ * Class for model class Athlete
+ */
 
 @Entity
 @Table(name = "athletes")
