+++ src/main/java/tapp/org/tapp/models/Tenant.java
@@ -1,9 +1,19 @@
-package tapp.org.tapp.Models;
+package tapp.org.tapp.models;
 
 
-import jakarta.persistence.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Lob;
+import jakarta.persistence.Table;
 
 import java.util.Objects;
+
+/**
+ * Class for model class Tenant. A tenant (or club) can have athletes and skills.
+ */
 
 @Entity
 @Table(name = "tenants")
