+++ src/main/java/tapp/org/tapp/models/TenantSkillID.java
@@ -1,10 +1,10 @@
-package tapp.org.tapp.Models;
+package tapp.org.tapp.models;
 
 import java.io.Serializable;
 import java.util.Objects;
 
 /**
- * Class is needed for the Progress Class which uses a composite primary key.
+ * Class is needed for the TenantSkill class which uses a composite primary key.
  */
 
 public  class TenantSkillID implements Serializable {
