+++ src/main/java/tapp/org/tapp/repository/ProgressRepository.java
@@ -1,8 +1,8 @@
-package tapp.org.tapp.Repository;
+package tapp.org.tapp.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import tapp.org.tapp.Models.Progress;
+import tapp.org.tapp.models.Progress;
 
 import java.util.List;
