+++ src/main/java/tapp/org/tapp/repository/RelatedSkillRepository.java
@@ -1,17 +1,13 @@
-package tapp.org.tapp.Repository;
+package tapp.org.tapp.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import tapp.org.tapp.Models.RelatedSkill;
+import tapp.org.tapp.models.RelatedSkill;
 
 import java.util.List;
 
 
 @Repository
 public interface RelatedSkillRepository extends JpaRepository<RelatedSkill, Long> {
-	//List<RelatedSkill> findAllrequiredskillByskillId(long skillId);
 	List<RelatedSkillsId> findAllrequiredskillByskillId(long skillId);
-
-
-
 }
