+++ src/main/java/tapp/org/tapp/repository/RelatedSkillsId.java
@@ -1,4 +1,4 @@
-package tapp.org.tapp.Repository;
+package tapp.org.tapp.repository;
 
 public interface RelatedSkillsId {
 	long getSkillId();
