+++ src/main/java/tapp/org/tapp/repository/TenantSkillRepository.java
@@ -1,8 +1,8 @@
-package tapp.org.tapp.Repository;
+package tapp.org.tapp.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import tapp.org.tapp.Models.TenantSkill;
+import tapp.org.tapp.models.TenantSkill;
 
 import java.util.List;
